Checking for existence of indexes on a table before dropping.

+++ database/migrations/2021_10_25_062848_user_delete_foreign_key_invite_id.php
@@ -15,9 +15,20 @@ class UserDeleteForeignKeyInviteId extends Migration
     {
         Schema::table('users', function (Blueprint $table) {
             $table->dropForeign('users_department_id_foreign');
-            $table->dropIndex('users_invite_id_foreign_custom');
-            $table->dropIndex('users_department_id_foreign');
-            $table->dropColumn('invite_id');
+
+            $sm = Schema::getConnection()->getDoctrineSchemaManager();
+            $doctrineTable = $sm->listTableDetails('users');
+
+            if ($doctrineTable->hasIndex('users_invite_id_foreign_custom')) {
+                $table->dropIndex('users_invite_id_foreign_custom');
+            }
+            if ($doctrineTable->hasIndex('users_department_id_foreign')) {
+                $table->dropIndex('users_department_id_foreign');
+            }
+
+            if (Schema::hasColumn('users', 'invite_id')) {
+                $table->dropColumn('invite_id');
+            }
         });
     }
